Update to session system
Testing revealed that in previous versions of my session implementation, a glitch was caused by more than one instance of a user account attempting to log in at one time. This update fixes this by adding a dedicated sessions database.

A later update will be required to cull old sessions.