Work on Upload Page and styling work
This is an interim update that primarily features changes to the upload page. It is still not complete, but more of the styling and functionality is in place.

Some work has been done to the nav bar but it didn't go very well and thus will all be undone.

The cropper library has been fixed. The code in it previously was not correct.

New photo svg icon added, new futura font, and the existing Vitae logo had its 'title' removed because it was showing up as a tooltip when hovered.

That is all.